feat(ai): complete V2 enhanced following algorithm with modular architecture
• Implement complete trump lead algorithm with unified beating thresholds
  - Dynamic strength thresholds: 150 for high points (≥10), 100 for low points (<10)
  - Strategic teammate contribution analysis with 4th player and trump strength checks
  - Intelligent takeover logic for both opponent and weak teammate scenarios
  - Conservative disposal fallback with strategic value optimization

• Create specialized validCombosDecision module for clean separation
  - handleTrumpLeadValidCombos() with sophisticated beating logic
  - handleNonTrumpLeadValidCombos() with memory-enhanced contribution analysis
  - Unified helper functions for contribution analysis and beating thresholds
  - Integrated combo selection utilities for cohesive functionality

• Enhance routingLogic with modular architecture
  - Clean scenario routing with focused responsibility
  - Reduced complexity by extracting specialized algorithms
  - Improved maintainability with clear module boundaries
  - Added combo requirements analysis for proper following validation

• Consolidate combo selection utilities into validCombosDecision
  - Merge comboSelection.ts into primary decision module
  - Reduce file count while maintaining logical cohesion
  - Eliminate cross-module dependencies for basic selection functions
  - Clean up import dependencies across the system

Phase 1 implementation now 90% complete with sophisticated modular architecture
supporting trump/non-trump scenarios, memory integration, and strategic optimization.
