+++ src/ai/followingV2/routingLogic.ts
@@ -1,10 +1,7 @@
-import { canBeatCombo } from "../../game/cardComparison";
 import { calculateCardStrategicValue, isTrump } from "../../game/cardValue";
-import { Card, GameContext, GameState, PlayerId, TrumpInfo } from "../../types";
+import { Card, ComboType, GameContext, GameState, PlayerId, TrumpInfo } from "../../types";
 import { gameLogger } from "../../utils/gameLogger";
-import { getRemainingUnseenCards } from "../aiGameContext";
 import {
-  analyzeSuitAvailability,
   SuitAvailabilityResult,
 } from "./suitAvailabilityAnalysis";
 // Memory integration V2 focuses on three core responsibilities:
@@ -12,12 +9,15 @@ import {
 // 2. Beatability analysis
 // 3. Remaining points analysis
 // These are used by strategic selection functions
-import { selectComboByStrategicValue } from "./comboSelection";
 import {
   selectLowestValueNonPointCombo,
   selectPointContribution,
   selectStrategicDisposal,
 } from "./strategicSelection";
+import {
+  handleTrumpLeadValidCombos,
+  handleNonTrumpLeadValidCombos,
+} from "./validCombosDecision";
 
 /**
  * Routing Logic - Main decision router for enhanced following strategy
@@ -119,177 +119,7 @@ export function routeToDecision(
   }
 }
 
-/**
- * Handle trump lead valid combos - specialized for trump competition
- *
- * V2: Enhanced algorithm with A/B/C/D strategy classification
- */
-function handleTrumpLeadValidCombos(
-  analysis: SuitAvailabilityResult,
-  context: GameContext,
-  trumpInfo: TrumpInfo,
-  gameState: GameState,
-  currentPlayerId: PlayerId,
-): Card[] {
-  if (!analysis.validCombos || analysis.validCombos.length === 0) {
-    gameLogger.error("enhanced_following_no_trump_valid_combos", {
-      player: currentPlayerId,
-      scenario: analysis.scenario,
-      message:
-        "handleTrumpLeadValidCombos called with no valid combos - routing bug",
-    });
-    return fallbackSelection([], context, gameState, trumpInfo);
-  }
 
-  gameLogger.debug("enhanced_following_trump_lead_valid_combos", {
-    player: currentPlayerId,
-    validComboCount: analysis.validCombos.length,
-    comboTypes: analysis.validCombos.map((c) => c.type),
-  });
-
-  // Shortcut: If only one valid combo, just play it
-  if (analysis.validCombos.length === 1) {
-    gameLogger.debug("enhanced_following_trump_single_combo_shortcut", {
-      player: currentPlayerId,
-      selectedCards: analysis.validCombos[0].cards.map((c) => c.toString()),
-      reason: "only_one_valid_combo",
-    });
-    return analysis.validCombos[0].cards;
-  }
-
-  // TODO: Implement V2 enhanced trump lead algorithm
-  // For now, use strategic disposal as fallback
-  const selectedCards = selectStrategicDisposal(
-    analysis.validCombos[0].cards,
-    analysis.requiredLength,
-    trumpInfo,
-  );
-
-  gameLogger.debug("enhanced_following_trump_lead_result", {
-    player: currentPlayerId,
-    selectedCardCount: selectedCards.length,
-    selectedCards: selectedCards.map((c) => c.toString()),
-  });
-
-  return selectedCards;
-}
-
-/**
- * Handle non-trump lead valid combos - specialized for non-trump situations
- *
- * Enhanced algorithm:
- * 1. Contribute if teammate secure + safe timing
- * 2. Beat if can beat + not trumped by opponent
- * 3. Dispose smallest combo avoiding points
- */
-function handleNonTrumpLeadValidCombos(
-  analysis: SuitAvailabilityResult,
-  context: GameContext,
-  trumpInfo: TrumpInfo,
-  gameState: GameState,
-  currentPlayerId: PlayerId,
-): Card[] {
-  if (!analysis.validCombos || analysis.validCombos.length === 0) {
-    gameLogger.error("enhanced_following_no_nontrump_valid_combos", {
-      player: currentPlayerId,
-      scenario: analysis.scenario,
-      message:
-        "handleNonTrumpLeadValidCombos called with no valid combos - routing bug",
-    });
-    return fallbackSelection([], context, gameState, trumpInfo);
-  }
-
-  const trickAnalysis = context.trickWinnerAnalysis;
-  if (!trickAnalysis) {
-    gameLogger.error("enhanced_following_no_trick_analysis", {
-      player: currentPlayerId,
-      message:
-        "handleNonTrumpLeadValidCombos called without trick analysis - context bug",
-    });
-    return analysis.validCombos[0].cards;
-  }
-
-  const { isTeammateWinning, isCurrentlyTrumped } = trickAnalysis;
-
-  gameLogger.debug("enhanced_following_nontrump_lead_analysis", {
-    player: currentPlayerId,
-    validComboCount: analysis.validCombos.length,
-    isTeammateWinning,
-    isCurrentlyTrumped,
-    comboTypes: analysis.validCombos.map((c) => c.type),
-  });
-
-  // Shortcut: If only one valid combo, just play it
-  if (analysis.validCombos.length === 1) {
-    gameLogger.debug("enhanced_following_single_combo_shortcut", {
-      player: currentPlayerId,
-      selectedCards: analysis.validCombos[0].cards.map((c) => c.toString()),
-      reason: "only_one_valid_combo",
-    });
-    return analysis.validCombos[0].cards;
-  }
-
-  // Step 1: Check if we should contribute to teammate
-  if (shouldContributeToTeammate(context, gameState, currentPlayerId)) {
-    const contributionCards = selectComboByStrategicValue(
-      analysis.validCombos,
-      trumpInfo,
-      "contribute",
-      "lowest",
-    );
-    if (contributionCards.length > 0) {
-      gameLogger.debug("enhanced_following_contribute_to_teammate", {
-        player: currentPlayerId,
-        selectedCards: contributionCards.map((c) => c.toString()),
-        reason: "teammate_secure_and_safe_timing",
-      });
-      return contributionCards;
-    }
-  }
-
-  // Step 2: Check if we should beat opponent (can beat + not trumped by opponent)
-  if (!isTeammateWinning && !isCurrentlyTrumped) {
-    const highestCombo = selectComboByStrategicValue(
-      analysis.validCombos,
-      trumpInfo,
-      "basic",
-      "highest",
-    );
-
-    // Check if this combo can actually beat the current winner
-    const currentWinnerCards =
-      gameState.currentTrick?.plays.find(
-        (play) => play.playerId === context.trickWinnerAnalysis?.currentWinner,
-      )?.cards || [];
-
-    const canBeat = canBeatCombo(highestCombo, currentWinnerCards, trumpInfo);
-
-    if (highestCombo.length > 0 && canBeat) {
-      gameLogger.debug("enhanced_following_beat_opponent", {
-        player: currentPlayerId,
-        selectedCards: highestCombo.map((c) => c.toString()),
-        reason: "can_beat_and_not_trumped",
-      });
-      return highestCombo;
-    }
-  }
-
-  // Step 3: Fallback - dispose smallest combo avoiding points
-  const disposalCards = selectComboByStrategicValue(
-    analysis.validCombos,
-    trumpInfo,
-    "strategic",
-    "lowest",
-  );
-
-  gameLogger.debug("enhanced_following_dispose_nontrump", {
-    player: currentPlayerId,
-    selectedCards: disposalCards.map((c) => c.toString()),
-    reason: "fallback_disposal",
-  });
-
-  return disposalCards;
-}
 
 /**
  * Handle enough remaining scenario - same suit contribute or dispose
@@ -624,132 +454,43 @@ function fallbackSelection(
   }
 }
 
-// =============== HELPER FUNCTIONS FOR NON-TRUMP LEAD ALGORITHM ===============
-
 /**
- * Determine if we should contribute points to teammate
- *
- * Conditions:
- * - teammate trumped it OR
- * - teammate winning non-trump AND biggest in suit by memory
- * AND
- * - (I'm 4th player OR next player not void)
+ * Get combo requirements based on leading combo type
  */
-function shouldContributeToTeammate(
-  context: GameContext,
-  gameState: GameState,
-  _currentPlayerId: PlayerId,
-): boolean {
-  const trickAnalysis = context.trickWinnerAnalysis;
-  if (!trickAnalysis?.isTeammateWinning) {
-    return false;
+function getComboRequirements(
+  leadingComboType: ComboType,
+  requiredLength: number,
+): {
+  minimumPairs: number;
+  minimumTractors: number;
+  canUseMixed: boolean;
+} {
+  switch (leadingComboType) {
+    case ComboType.Single:
+      return {
+        minimumPairs: 0,
+        minimumTractors: 0,
+        canUseMixed: true,
+      };
+    case ComboType.Pair:
+      return {
+        minimumPairs: requiredLength / 2, // Each pair is 2 cards
+        minimumTractors: 0,
+        canUseMixed: false,
+      };
+    case ComboType.Tractor:
+      return {
+        minimumPairs: requiredLength / 2, // Each pair is 2 cards
+        minimumTractors: 1, // Need at least one tractor
+        canUseMixed: false,
+      };
+    default:
+      // Multi-combo or other - treat as singles
+      return {
+        minimumPairs: 0,
+        minimumTractors: 0,
+        canUseMixed: true,
+      };
   }
-
-  const currentTrick = gameState.currentTrick;
-  if (!currentTrick) return false;
-
-  // Check if I'm 4th player
-  const isLastPlayer = currentTrick.plays.length === 3;
-
-  // Check next player void status (only if not last player)
-  let isNextPlayerVoid = false;
-  if (!isLastPlayer) {
-    const nextPlayerIndex = (gameState.currentPlayerIndex + 1) % 4;
-    const nextPlayerId = gameState.players[nextPlayerIndex]?.id as PlayerId;
-    const leadingSuit = currentTrick.plays[0]?.cards[0]?.suit;
-
-    if (context.memoryContext?.cardMemory && nextPlayerId && leadingSuit) {
-      const nextPlayerMemory =
-        context.memoryContext.cardMemory.playerMemories[nextPlayerId];
-      isNextPlayerVoid = nextPlayerMemory?.suitVoids.has(leadingSuit) ?? false;
-    }
-  }
-
-  // Condition 1: Teammate trumped it
-  if (trickAnalysis.isCurrentlyTrumped) {
-    return isLastPlayer || !isNextPlayerVoid;
-  }
-
-  // Condition 2: Teammate winning non-trump AND biggest in suit by memory
-  const currentWinnerPlay = currentTrick.plays.find(
-    (play) => play.playerId === context.trickWinnerAnalysis?.currentWinner,
-  );
-  const teammateIsBiggestInSuit = currentWinnerPlay
-    ? checkComboIsBiggestInSuit(
-        currentWinnerPlay.cards,
-        context,
-        gameState,
-        currentTrick,
-      )
-    : true;
-
-  if (teammateIsBiggestInSuit) {
-    return isLastPlayer || !isNextPlayerVoid;
-  }
-
-  return false;
 }
 
-/**
- * Check if given combo is biggest in the leading suit by memory analysis
- *
- * Algorithm:
- * 1. Get all remaining unseen cards for the leading suit
- * 2. Use suit availability analysis to see what combos are possible from unseen cards
- * 3. Check if any of those combos can beat the given combo
- * 4. If yes, combo is not biggest; if no, combo is biggest
- */
-function checkComboIsBiggestInSuit(
-  combo: Card[],
-  context: GameContext,
-  gameState: GameState,
-  currentTrick: NonNullable<GameState["currentTrick"]>,
-): boolean {
-  const leadingSuit = currentTrick.plays[0]?.cards[0]?.suit;
-  const trumpInfo = context.trumpInfo || gameState.trumpInfo;
-
-  if (!leadingSuit || !trumpInfo) {
-    return true; // Assume combo is biggest if we can't analyze
-  }
-
-  // Skip analysis if this is trump lead (trump cards don't follow this logic)
-  if (currentTrick.plays[0]?.cards.some((card) => isTrump(card, trumpInfo))) {
-    return true;
-  }
-
-  // Get all remaining unseen cards for the leading suit
-  const unseenCards = getRemainingUnseenCards(leadingSuit, context, gameState);
-
-  if (unseenCards.length === 0) {
-    return true; // No more cards in suit, combo definitely is biggest
-  }
-
-  // Use suit availability analysis to see what combos are possible from unseen cards
-  const leadingCards = currentTrick.plays[0]?.cards || [];
-  const unseenAnalysis = analyzeSuitAvailability(
-    leadingCards,
-    unseenCards, // Treat unseen cards as a hypothetical player's hand
-    trumpInfo,
-  );
-
-  // If no valid combos can be formed from unseen cards, combo is biggest
-  if (
-    unseenAnalysis.scenario !== "valid_combos" ||
-    !unseenAnalysis.validCombos ||
-    unseenAnalysis.validCombos.length === 0
-  ) {
-    return true;
-  }
-
-  // Check if any unseen combo can beat the given combo
-  for (const unseenCombo of unseenAnalysis.validCombos) {
-    const canBeat = canBeatCombo(unseenCombo.cards, combo, trumpInfo);
-    if (canBeat) {
-      // Found an unseen combo that can beat given combo - it is not biggest
-      return false;
-    }
-  }
-
-  // No unseen combo can beat given combo - it is biggest in suit
-  return true;
-}
